requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
diff-match-patch==20241021
Django==4.2.16
django-bootstrap-v5==1.0.11
django-import-export==4.2.1
pillow==11.0.0
soupsieve==2.6
sqlparse==0.5.1
tablib==3.7.0
tzdata==2024.2
whitenoise==6.8.2
